feat: pss-cln-def-fold-snapshots-002 - Fold replayconfig implementation into snapshots_portability
