+++ pkg/services/factory_definitions/internal/services/snapshots_portability/boundary_test.go
@@ -32,6 +32,7 @@ var snapshotsPortabilityForbiddenImportRoots = []string{
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/snapshotcapture",
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/portableconfig",
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/editable",
+	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/replayconfig",
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/definition",
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/loading",
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/persistence",
